requirements.txt
ddddocr==1.4.11
pycryptodome==3.20.0
pydantic==2.6.2
python-dotenv==1.0.1
Requests==2.31.0
rsa==4.9
ujson==5.9.0
